refactor: split vue file
